change parameter type in some methods in ExceptionUtility from "int" to "NSInteger"
reason for the change: callers may want to supply an UIKit enumeration
value, and UIKit enumerations are typically typed as NSUInteger or
NSInteger. these in turn are typedefs for "unsigned long" or "long",
so don't fit into an "int".

+++ src/utility/ExceptionUtility.h
@@ -28,10 +28,10 @@
 
 + (void) throwInvalidUIType:(enum UIType)uiType;
 + (void) throwInvalidArgumentExceptionWithFormat:(NSString*)format
-                                   argumentValue:(int)argumentValue;
+                                   argumentValue:(NSInteger)argumentValue;
 + (void) throwInvalidArgumentExceptionWithErrorMessage:(NSString*)errorMessage;
 + (void) throwInternalInconsistencyExceptionWithFormat:(NSString*)format
-                                         argumentValue:(int)argumentValue;
+                                         argumentValue:(NSInteger)argumentValue;
 + (void) throwInternalInconsistencyExceptionWithErrorMessage:(NSString*)errorMessage;
 + (void) throwAbstractMethodException;
 + (void) throwNotImplementedException;
